pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.5.0", "pybind11>=2.10"]
build-backend = "scikit_build_core.build"

[project]
name = "dftcu"
version = "0.1.0"
description = "CUDA-accelerated DFT calculations compatible with DFTpy"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["DFT", "CUDA", "computational-chemistry", "GPU"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: C++",
]

dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "ase>=3.22.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "click>=8.0.0",
]

[project.scripts]
dftcu = "dftcu.cli.main:main"

[tool.scikit-build]
cmake.version = ">=3.18"
# Use a shared build directory with manual CMake builds
# This allows sharing build artifacts between `cmake --build` and `uv pip install`
build-dir = "build"
cmake.args = [
    "-DBUILD_TESTING=OFF",
    "-DBUILD_DOCS=OFF",
    "-DCMAKE_COMPILE_WARNING_AS_ERROR=OFF"
]
# Configure editable install for incremental builds
# Use 'redirect' mode - faster than 'inplace' but requires manual rebuild
editable.mode = "redirect"
editable.rebuild = true
editable.verbose = true
# Install from existing build if available
install.strip = false
# Removed wheel.py-api = "cp39" to use current Python ABI for best performance


[project.optional-dependencies]
dev = [
    # Build system (Crucial for --no-build-isolation)
    "scikit-build-core>=0.5.0",
    "pybind11>=2.10",
    "cmake>=3.18",
    "ninja",  # 强烈推荐：加快编译速度

    # Testing
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "pytest-benchmark>=4.0.0",

    # Code quality
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",

    # Pre-commit hooks
    "pre-commit>=3.0.0",
]

docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.0.0",
]
benchmark = [
    "matplotlib>=3.5.0",
    "pandas>=1.5.0",
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
]
all = [
    "dftcu[dev,docs,benchmark]",
]

[project.urls]
Homepage = "https://github.com/your-org/DFTcu"
Documentation = "https://dftcu.readthedocs.io"
Repository = "https://github.com/your-org/DFTcu"
Issues = "https://github.com/your-org/DFTcu/issues"

# Tool configurations
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.venv
  | _build
  | build
  | dist
  | external
  | tests/nscf_alignment
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
extend_skip = ["external", "tests/nscf_alignment"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
norecursedirs = ["external", "build*", ".venv", ".git", "docs"]

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/external/*",
    "*/.venv/*",
    "*/build/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "external/",
    "build/",
    ".venv/",
    "tests/nscf_alignment/",
]
